changes settings, moves files and gets rid of eslint errors so the project will build
