+++ .eslintrc.json
@@ -1,7 +1,14 @@
 {
+  "plugins": ["jest-dom"],
+  "rules": {
+    "jest-dom/prefer-checked": "error",
+    "jest-dom/prefer-enabled-disabled": "error",
+    "jest-dom/prefer-required": "error",
+    "jest-dom/prefer-to-have-attribute": "error"
+  },
   "extends": [
     "next/core-web-vitals", 
     "plugin:testing-library/react",
-    "plugin:jest-dom/recommend"
+    "prettier"
   ]
 }
