+++ app/actions/getCurrentUser.ts
@@ -1,6 +1,6 @@
 import { getServerSession } from "next-auth/next";
 import { Session } from "next-auth";
-import { authOptions } from "../api/auth/[...nextauth]/route";
+import { authOptions } from "@/app/libs/configs/auth/authOptions";
 import prisma from "@/app/libs/prismadb";
 
 export async function getSession(): Promise<Session | null | undefined> {
