+++ app/components/modals/__tests__/heading.test.tsx
@@ -1,9 +1,6 @@
-import { render, screen, cleanup } from "@testing-library/react";
+import { render, screen } from "@testing-library/react";
 import Heading from "../Heading";
 
-afterEach(() => {
-	cleanup();
-});
 
 describe("Heading", () => {
 	describe("Render", () => {
