+++ app/listings/[listingId]/ListingClient.tsx
@@ -67,7 +67,7 @@ const ListingClient: React.FC<ListingClientProps> = ({ reservations = [], listin
 			})
 			.catch(() => toast.error("Something went wrong"))
 			.finally(() => setIsLoading(false));
-	}, [totalPrice, dateRange, listing.id, router, currentUser, loginModal]);
+	}, [dateRange, listing.id, router, currentUser, loginModal]);
 
 	useEffect(() => {
 		if (dateRange.startDate && dateRange.endDate && listing.price) {
